requirements.txt
torch 
torchvision
faiss-cpu
numpy 
pandas
matplotlib
scikit-learn
gradio
tqdm
Pillow
jupyter